+++ projects/Mallard/src/components/front/PageAppearance.ts
@@ -1,0 +1,8 @@
+export enum PageAppearance {
+    superhero,
+    two,
+    three,
+    four,
+    five,
+    six,
+}
